src/KLTN-AI/recommendations/requirements.txt
pandas
numpy
scikit-learn
fastapi
uvicorn
python-dotenv
sqlalchemy
plotly
pymysql
redis
implicit
apscheduler
tqdm
threadpoolctl